# griffin-hub/openapi-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "griffin-runner API",
    "description": "API for managing and running test plans",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {},
...
  },
  "paths": {
...
    "/agents/register": {
      "post": {
        "tags": ["agents"],
        "description": "Register a new agent with the hub",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["location"],
                "properties": {
                  "location": {
                    "type": "string"
                  },
                  "metadata": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "location",
                    "status",
                    "lastHeartbeat",
                    "registeredAt"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "location": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": ["online", "offline"],
                      "title": "AgentStatus"
                    },
                    "lastHeartbeat": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "registeredAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "metadata": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/agents/{id}/heartbeat": {
      "post": {
        "tags": ["agents"],
        "description": "Record a heartbeat from an agent",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Agent ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["success"],
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/agents/{id}": {
      "delete": {
        "tags": ["agents"],
        "description": "Deregister an agent",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Agent ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["success"],
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/agents/": {
      "get": {
        "tags": ["agents"],
        "description": "List all agents with optional filtering",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "location",
            "required": false,
            "description": "Filter by location"
          },
          {
            "schema": {
              "type": "string",
              "enum": ["online", "offline"]
            },
            "in": "query",
            "name": "status",
            "required": false,
            "description": "Filter by status"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["data", "total"],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "id",
                          "location",
                          "status",
                          "lastHeartbeat",
                          "registeredAt"
                        ],
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "location": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string",
                            "enum": ["online", "offline"]
                          },
                          "lastHeartbeat": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "registeredAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "metadata": {
                            "type": "object",
                            "additionalProperties": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    },
                    "total": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
}